backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
sqlalchemy>=2.0,<3
alembic>=1.13,<2
asyncpg>=0.29,<1
redis>=5.0,<6
rq>=1.16,<2
minio>=7.2,<8
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
python-multipart>=0.0.9
Pillow>=10.0

# dev
pytest>=8.0,<9
httpx>=0.27,<1
pytest-asyncio>=0.23,<1
aiosqlite>=0.20
